+++ R/utils.R
@@ -1,8 +1,14 @@
 #' @importFrom utils head
 load_object <- function(file) {
+  if (!grepl("\\.(rda|RData)$", file, ignore.case = TRUE)) {
+    usethis::ui_stop("{usethis::ui_path(file)} is not an .rda file. data/ holds one .rda file per data object, as usethis::use_data() writes them.")
+  }
   tmp_env <- new.env()
-  load(file = file, envir = tmp_env)
-  tmp_env[[ls(tmp_env)[1]]]
+  loaded <- load(file = file, envir = tmp_env)
+  if (length(loaded) != 1) {
+    usethis::ui_stop("{usethis::ui_path(file)} holds {length(loaded)} objects ({usethis::ui_value(loaded)}). washr expects one data object per .rda file, as usethis::use_data() writes them.")
+  }
+  tmp_env[[loaded]]
 }
 
 is_pkg <- function(){